finished first task

+++ FVCalc.java
@@ -1,7 +1,7 @@
 // Computes the future value of a saving investment.
 public class FVCalc {
 	public static void main(String[] args){
-		double currentValue = Double.parseDouble(args[0]);
+		int currentValue = Integer.parseInt(args[0]);
 		double interestRate = ( Double.parseDouble(args[1]));
 		int year = Integer.parseInt(args[2]);
 		int futureValue = (int)((currentValue * (Math.pow (1+ (interestRate/100), year))));
